Revert "Default boottidal path"
