+++ lib/repl.js
@@ -2,7 +2,7 @@
 
 var fs = require('fs');
 const Ghci = require('./ghci')
-const BootTidal = require('./boot-tidal')
+var defaultBootFileName = 'BootTidal.hs';
 var CONST_LINE = 'line'
 var CONST_MULTI_LINE = 'multi_line'
 
@@ -102,25 +102,33 @@ export default class REPL {
 
     }
 
-    initTidal() {
-        const bootPath = BootTidal.getPath()
-        this.consoleView.logStdout(`Load BootTidal.hs from ${bootPath}`)
-        try {
-          const blocks = fs.readFileSync(bootPath)
-              .toString()
-              .split('\n\n')
-              .map(block => block.replace(":{", "").replace(":}", ""));
+    getBootTidalPath(rootDirectories) {
+        const configuredBootFilePath = atom.config.get('tidalcycles.bootTidalPath');
+        if (configuredBootFilePath) return configuredBootFilePath
 
-          blocks.forEach(block => {
-              if (block.startsWith(":set")) {
-                  block.split("\n").forEach(row => this.tidalSendLine(row))
-              } else {
-                  this.tidalSendExpression(block);
-              }
-          });
-        } catch (err) {
-          this.consoleView.logStderr(`Error initializing Tidal: ${err}`)
-        }
+        const currentDirectoryPath = rootDirectories.length > 0 ?
+            rootDirectories[0].path + '/' + defaultBootFileName :
+            null;
+
+        if (fs.existsSync(currentDirectoryPath)) return currentDirectoryPath;
+
+        return  __dirname + '/' + defaultBootFileName;
+    }
+
+    initTidal() {
+        const bootPath = this.getBootTidalPath(atom.project.rootDirectories);
+        const blocks = fs.readFileSync(bootPath)
+            .toString()
+            .split('\n\n')
+            .map(block => block.replace(":{", "").replace(":}", ""));
+
+        blocks.forEach(block => {
+            if (block.startsWith(":set")) {
+                block.split("\n").forEach(row => this.tidalSendLine(row))
+            } else {
+                this.tidalSendExpression(block);
+            }
+        });
     }
 
     tidalSendExpression(expression) {
@@ -139,14 +147,15 @@ export default class REPL {
     start() {
       this.consoleView.initUI();
       this.ghci = Ghci.spawn()
-        .on('stderr', data => { this.processStdErr(data) })
+        .on('stderr', (data) => { this.processStdErr(data) })
         .on('stdout', data => { this.processStdOut(data) })
 
       this.initTidal();
     }
 
     getEditor() {
-        return atom.workspace.getActiveTextEditor();
+        var editor = atom.workspace.getActiveTextEditor();
+        return editor;
     }
 
     eval(evalType, copy) {
