--- spec/boot-tidal-spec.js
@@ -1,52 +1,0 @@
-const BootTidal = require('../lib/boot-tidal.js')
-const fs = require('fs')
-const child_process = require('child_process')
-
-describe('boot-tidal', () => {
-  let bootTidal = new BootTidal([{ path: '/current/directory' }])
-
-  beforeEach(() => {
-    this.originalPlatform = process.platform;
-    waitsForPromise(() => atom.packages.activate('tidalcycles'))
-  })
-
-  afterEach(() => {
-    setPlatform(this.originalPlatform)
-  })
-
-  describe('boot file sequence', () => {
-
-    it(`should choose current directory BootTidal.hs if it exists`, () => {
-      spyOn(fs, 'existsSync').andReturn(true)
-
-      expect(bootTidal.choosePath()).toBe('/current/directory/BootTidal.hs')
-    })
-
-    it(`should choose custom boot file when provided`, () => {
-      atom.config.set('tidalcycles.bootTidalPath', '/custom/directory/BootTidal.hs')
-      spyOn(fs, 'existsSync').andReturn(true)
-
-      expect(bootTidal.choosePath()).toBe('/custom/directory/BootTidal.hs')
-    })
-
-    it('should choose tidal installation boot file when OS is a known one', () => {
-      spyOn(fs, 'existsSync').andReturn(false)
-      spyOn(child_process, 'execSync').andReturn('/posix/tidal\n')
-      setPlatform('linux')
-
-      expect(bootTidal.choosePath()).toBe('/posix/tidal/BootTidal.hs')
-    })
-
-    it(`should choose default boot file when no custom provided, no file in current directory and platform is unknown`, () => {
-      spyOn(fs, 'existsSync').andReturn(false)
-      setPlatform('unknown')
-
-      expect(bootTidal.choosePath()).toContain('/lib/BootTidal.hs')
-    })
-
-  })
-})
-
-function setPlatform(name) {
-  Object.defineProperty(process, 'platform', { value: name });
-}
