+++ spec/repl-spec.js
@@ -1,10 +1,8 @@
 const REPL = require('../lib/repl.js')
-const BootTidal = require('../lib/boot-tidal')
 const fs = require('fs')
 
 describe('repl', () => {
-  let consoleView = { logStdout: () => {} }
-  let repl = new REPL(consoleView)
+  let repl = new REPL()
 
   beforeEach(() => {
     waitsForPromise(() => atom.packages.activate('tidalcycles'))
@@ -13,7 +11,7 @@ describe('repl', () => {
   describe('init tidal', () => {
 
     it('should load code blocks on boot', () => {
-      spyOn(BootTidal, 'getPath').andReturn("lib/BootTidal.hs")
+      spyOn(repl, 'getBootTidalPath').andReturn("lib/BootTidal.hs")
       spyOn(repl, 'tidalSendLine')
       spyOn(repl, 'tidalSendExpression')
 
@@ -26,7 +24,7 @@ describe('repl', () => {
     })
 
     it('should send :set lines as single lines on boot', () => {
-      spyOn(BootTidal, 'getPath').andReturn("lib/BootTidal.hs")
+      spyOn(repl, 'getBootTidalPath').andReturn("lib/BootTidal.hs")
       spyOn(repl, 'tidalSendLine')
       spyOn(repl, 'tidalSendExpression')
 
@@ -36,6 +34,29 @@ describe('repl', () => {
       expect(repl.tidalSendLine.calls[1].args[0]).toBe(':set prompt ""')
       expect(repl.tidalSendLine.calls[2].args[0]).toBe(':set prompt-cont ""')
       expect(repl.tidalSendLine.calls[3].args[0]).toBe(':set prompt "tidal> "')
+    })
+  })
+
+  describe('boot file sequence', () => {
+    let rootDirectories = [{ path: '/current/directory' }]
+
+    it(`should choose current directory BootTidal.hs if it exists`, () => {
+      spyOn(fs, 'existsSync').andReturn(true)
+
+      expect(repl.getBootTidalPath(rootDirectories)).toBe('/current/directory/BootTidal.hs')
+    })
+
+    it(`should choose custom boot file when provided`, () => {
+      atom.config.set('tidalcycles.bootTidalPath', '/custom/directory/BootTidal.hs')
+      spyOn(fs, 'existsSync').andReturn(true)
+
+      expect(repl.getBootTidalPath(rootDirectories)).toBe('/custom/directory/BootTidal.hs')
+    })
+
+    it(`should choose default boot file when no custom provided and no file in current directory`, () => {
+      spyOn(fs, 'existsSync').andReturn(false)
+
+      expect(repl.getBootTidalPath([])).toContain('/lib/BootTidal.hs')
     })
   })
 })
